Changed name of the file to SDWriter. Added it's own header file. May or may not record data properly. Or work at all.
