+++ include/SDWriter.h
@@ -1,0 +1,7 @@
+#ifndef SDWRITER_H
+#define SDWRITER_H
+
+void initSDCard();
+void SDCardWrite();
+
+#endif
